Return post data from post delete endpoint

+++ lib/class-wp-json-posts.php
@@ -369,18 +369,33 @@ class WP_JSON_Posts {
 			return new WP_Error( 'json_user_cannot_delete_post', __( 'Sorry, you are not allowed to delete this post.' ), array( 'status' => 401 ) );
 		}
 
+		// Grab the post first, so we can return the data later
+		$response = json_ensure_response( $this->get( $id, 'edit' ) );
+		if ( is_wp_error( $response ) ) {
+			return $response;
+		}
+
 		$result = wp_delete_post( $id, $force );
 
 		if ( ! $result ) {
 			return new WP_Error( 'json_cannot_delete', __( 'The post cannot be deleted.' ), array( 'status' => 500 ) );
 		}
 
-		if ( $force ) {
-			return array( 'message' => __( 'Permanently deleted post' ) );
-		} else {
-			// TODO: return a HTTP 202 here instead
-			return array( 'message' => __( 'Deleted post' ) );
+		// Update the response to indicate the new status
+		$data = $response->get_data();
+		$new_status = get_post_status( $id );
+		if ( $new_status === 'trash' ) {
+			// Send a 202 if it's only trashed to indicate that it will be
+			// actually deleted later
+			$response->set_status( 202 );
+			$data['status'] = 'trash';
 		}
+		else {
+			$data['status'] = 'deleted';
+		}
+		$response->set_data( $data );
+
+		return $response;
 	}
 
 	/**
